% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{AIC.oblr}
\alias{AIC.oblr}
\title{AIC Method for oblr Objects}
\usage{
\method{AIC}{oblr}(object, ..., k = 2)
}
\arguments{
\item{object}{An object of class \code{oblr}.}

\item{...}{Additional arguments passed to or from other methods.}

\item{k}{The penalty per parameter to be used in the AIC calculation. Default is 2.}
}
\value{
A numeric value representing the AIC.
}
\description{
Calculates the Akaike Information Criterion (AIC) for the \code{oblr} model.
}
