% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{anova.oblr}
\alias{anova.oblr}
\title{Anova Method for oblr Objects}
\usage{
\method{anova}{oblr}(object, ..., test = c("Chisq", "F", "none"))
}
\arguments{
\item{object}{An object of class "oblr", typically the result of a call to oblr().}

\item{...}{Additional objects of class "oblr", or a single object of class "list" containing only objects of class "oblr".}

\item{test}{A character string specifying the test statistic to be used.
Can be one of "Chisq" (default) for likelihood ratio test,
"F" for F-test, or "none" to skip significance testing.}
}
\value{
An object of class "anova" inheriting from class "data.frame".
}
\description{
This function performs an analysis of variance (or more precisely, an analysis of deviance)
for one or more fitted logistic regression model objects of class 'oblr'.
}
