% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_obwoe.R
\name{bake.step_obwoe}
\alias{bake.step_obwoe}
\title{Apply the Optimal Binning Transformation}
\usage{
\method{bake}{step_obwoe}(object, new_data, ...)
}
\arguments{
\item{object}{A trained step_obwoe object.}

\item{new_data}{A tibble or data frame to transform.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A tibble with transformed columns according to the \code{output}
  parameter.
}
\description{
Applies the learned binning and WoE transformation to new data. This method
is called by \code{\link[recipes]{bake}} and should not be invoked directly.
}
