% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{BIC.oblr}
\alias{BIC.oblr}
\title{BIC Method for oblr Objects}
\usage{
\method{BIC}{oblr}(object, ...)
}
\arguments{
\item{object}{An object of class \code{oblr}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A numeric value representing the BIC.
}
\description{
Calculates the Bayesian Information Criterion (BIC) for the \code{oblr} model.
}
