% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{CalculateSpecialWoE}
\alias{CalculateSpecialWoE}
\title{Calculate Special WoE}
\usage{
CalculateSpecialWoE(target)
}
\arguments{
\item{target}{Target values for special cases}
}
\value{
WoE value for special cases
}
\description{
Calculate Special WoE
}
\keyword{internal}
