% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{coef.oblr}
\alias{coef.oblr}
\title{Coefficients Method for oblr Objects}
\usage{
\method{coef}{oblr}(object, ...)
}
\arguments{
\item{object}{An object of class \code{oblr}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A numeric vector of estimated coefficients.
}
\description{
Extracts the estimated coefficients from the \code{oblr} model.
}
