% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{computeMetrics}
\alias{computeMetrics}
\title{Compute Performance Metrics for Logistic Regression Models}
\usage{
computeMetrics(object, newdata = NULL, cutoff = 0.5)
}
\arguments{
\item{object}{An object of class "oblr".}

\item{newdata}{A data frame or data.table containing new data for evaluation. If NULL, uses the data from the fit.}

\item{cutoff}{The probability cutoff for class prediction. Default is 0.5.}
}
\value{
A data.table with the calculated metrics.
}
\description{
Calculates various performance metrics for an oblr model.
}
\details{
This function calculates the following metrics:
\enumerate{
\item Log-likelihood (LogLik):
\deqn{LogLik = \sum_{i=1}^n [y_i \log(p_i) + (1-y_i) \log(1-p_i)]}
where \eqn{y_i} are the observed values and \eqn{p_i} are the predicted probabilities.
\item Akaike Information Criterion (AIC):
\deqn{AIC = 2k - 2LogLik}
where \eqn{k} is the number of parameters in the model.
\item Bayesian Information Criterion (BIC):
\deqn{BIC = k\log(n) - 2LogLik}
where \eqn{n} is the number of observations.
\item Area Under the ROC Curve (AUC):
AUC is the area under the Receiver Operating Characteristic curve, which plots the true positive rate against the false positive rate.
\item Gini Coefficient:
\deqn{Gini = 2 * AUC - 1}
\item Kolmogorov-Smirnov Statistic (KS):
\deqn{KS = \max|F_1(x) - F_0(x)|}
where \eqn{F_1(x)} and \eqn{F_0(x)} are the cumulative distribution functions for the positive and negative classes, respectively.
\item Accuracy:
\deqn{Accuracy = \frac{TP + TN}{TP + TN + FP + FN}}
where TP = True Positives, TN = True Negatives, FP = False Positives, FN = False Negatives.
\item Recall (Sensitivity):
\deqn{Recall = \frac{TP}{TP + FN}}
\item Precision:
\deqn{Precision = \frac{TP}{TP + FP}}
\item F1-Score:
\deqn{F1 = 2 * \frac{Precision * Recall}{Precision + Recall}}
}

These metrics provide a comprehensive view of the model's performance,
including its predictive capability (AUC, KS), fit to the data (LogLik, AIC, BIC),
and performance in classification tasks (Accuracy, Recall, Precision, F1-Score).
}
