% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{CreateSpecialBin}
\alias{CreateSpecialBin}
\title{Create Special Bin}
\usage{
CreateSpecialBin(dt_special, woebin, special_woe)
}
\arguments{
\item{dt_special}{Data for special cases}

\item{woebin}{Existing WoE bins}

\item{special_woe}{WoE value for special cases}
}
\value{
Special bin information
}
\description{
Create Special Bin
}
\keyword{internal}
