% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe_scorecard.R
\name{.ob_fit_checked}
\alias{.ob_fit_checked}
\title{Internal: Fit With Sign, Separation and Convergence Checks}
\usage{
.ob_fit_checked(eng, x, y, engine_args, control, add_warning)
}
\description{
Internal: Fit With Sign, Separation and Convergence Checks
}
\keyword{internal}
