% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe_scorecard.R
\name{.ob_score_sample}
\alias{.ob_score_sample}
\title{Internal: Score One Sample}
\usage{
.ob_score_sample(
  df,
  name,
  binning,
  features,
  eng,
  model,
  scaling,
  points,
  target,
  control,
  add_warning
)
}
\description{
Internal: Score One Sample
}
\keyword{internal}
