% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe_sql.R
\name{.ob_sql_reserved}
\alias{.ob_sql_reserved}
\title{Internal: Reserved Words Requiring Identifier Quoting}
\usage{
.ob_sql_reserved()
}
\description{
Internal: Reserved Words Requiring Identifier Quoting
}
\keyword{internal}
