% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_obwoe.R
\name{.universal_algorithms}
\alias{.universal_algorithms}
\title{Universal Algorithms}
\usage{
.universal_algorithms()
}
\value{
A character vector of universal algorithm names.
}
\description{
Internal function returning algorithm identifiers that support both
numerical and categorical features.
}
