% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{fitted.oblr}
\alias{fitted.oblr}
\title{Fitted Values Method for oblr Objects}
\usage{
\method{fitted}{oblr}(object, ...)
}
\arguments{
\item{object}{An object of class \code{oblr}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A numeric vector of fitted values.
}
\description{
Returns the fitted values (predicted probabilities) from the \code{oblr} model.
}
