% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{logLik.oblr}
\alias{logLik.oblr}
\title{Log-Likelihood Method for oblr Objects}
\usage{
\method{logLik}{oblr}(object, ...)
}
\arguments{
\item{object}{An object of class \code{oblr}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
An object of class \code{logLik} containing the log-likelihood.
}
\description{
Returns the log-likelihood of the \code{oblr} model.
}
