% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OBCreateSpecialBin}
\alias{OBCreateSpecialBin}
\title{Create Special Bin Entry}
\usage{
OBCreateSpecialBin(dt_special, woebin, special_woe)
}
\arguments{
\item{dt_special}{data.frame of special-case rows.}

\item{woebin}{data.frame of existing bins (must have count_pos, count_neg).}

\item{special_woe}{Numeric WoE value.}
}
\value{
Single-row data.frame with special bin info.
}
\description{
Create Special Bin Entry
}
