% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OBMapTargetVariable}
\alias{OBMapTargetVariable}
\title{Map Target Variable to 0/1}
\usage{
OBMapTargetVariable(dt, target, positive)
}
\arguments{
\item{dt}{A data.frame.}

\item{target}{Name of the target column.}

\item{positive}{Which level is positive, "bad|1" or "good|1".}
}
\value{
data.frame with recoded target.
}
\description{
Map Target Variable to 0/1
}
