% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OBSelectAlgorithm}
\alias{OBSelectAlgorithm}
\title{Select Optimal Binning Algorithm (Unused in main flow)}
\usage{
OBSelectAlgorithm(feature, method, dt, min_bin, max_bin, control)
}
\arguments{
\item{feature}{Name of the feature.}

\item{method}{Desired method.}

\item{dt}{data.frame.}

\item{min_bin}{Integer.}

\item{max_bin}{Integer.}

\item{control}{List.}
}
\value{
List(algorithm, params, method).
}
\description{
Select Optimal Binning Algorithm (Unused in main flow)
}
