% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OBSelectBestModel}
\alias{OBSelectBestModel}
\title{Select the Best Model for Optimal Binning}
\usage{
OBSelectBestModel(
  dt,
  target,
  features,
  method = NULL,
  min_bins,
  max_bins,
  control,
  progress = TRUE,
  trace = FALSE
)
}
\arguments{
\item{dt}{data.frame with target + feature columns.}

\item{target}{Name of the target column.}

\item{features}{Character vector of feature names.}

\item{method}{String or vector; methods to try or "auto".}

\item{min_bins}{Integer.}

\item{max_bins}{Integer.}

\item{control}{List of control parameters.}

\item{progress}{Logical; show progress bar.}

\item{trace}{Logical; print method errors.}
}
\value{
Named list of results per feature: each has woebin, woefeature, report, bestmethod.
}
\description{
Select the Best Model for Optimal Binning
}
