% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OBValidateInputs}
\alias{OBValidateInputs}
\title{Validate Inputs for Optimal Binning}
\usage{
OBValidateInputs(
  dt,
  target,
  features,
  method,
  preprocess,
  min_bins,
  max_bins,
  control,
  positive
)
}
\arguments{
\item{dt}{A data.frame.}

\item{target}{Target column name.}

\item{features}{Character vector or NULL.}

\item{method}{String or vector of methods.}

\item{preprocess}{Logical.}

\item{min_bins}{Integer.}

\item{max_bins}{Integer.}

\item{control}{List of control parameters.}

\item{positive}{"bad|1" or "good|1".}
}
\value{
NULL; stops on error.
}
\description{
Validate Inputs for Optimal Binning
}
