% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{obwoe_algorithms}
\alias{obwoe_algorithms}
\title{List Available Algorithms}
\usage{
obwoe_algorithms()
}
\value{
A data frame with algorithm information.
}
\description{
Returns a data frame with all available binning algorithms.
}
\examples{
obwoe_algorithms()

}
