% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OBWoEMonotonic}
\alias{OBWoEMonotonic}
\title{Check if WoE values are monotonic}
\usage{
OBWoEMonotonic(woe_values)
}
\arguments{
\item{woe_values}{Vector of WoE values}
}
\value{
Logical indicating if WoE values are monotonic
}
\description{
Check if WoE values are monotonic
}
\keyword{internal}
