% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimalBinningCategoricalIV}
\alias{OptimalBinningCategoricalIV}
\title{Performs optimal binning of a categorical variable for Weight of Evidence (WoE) and Information Value (IV), ensuring monotonicity and maximizing IV}
\usage{
OptimalBinningCategoricalIV(
  target,
  feature,
  min_bins = 2L,
  max_bins = 7L,
  cat_cutoff = 0.05,
  min_bads = 0.05
)
}
\arguments{
\item{target}{Integer vector representing the binary target variable, where 1 indicates a positive event (e.g., default) and 0 indicates a negative event (e.g., non-default).}

\item{feature}{Character vector representing the categorical variable to be binned.}

\item{min_bins}{(Optional) Minimum number of bins to generate. Default is 2.}

\item{max_bins}{(Optional) Maximum number of bins to generate. Default is 7.}

\item{cat_cutoff}{(Optional) Frequency cutoff value, below which categories are grouped into "Other". Default is 0.05.}

\item{min_bads}{(Optional) Minimum proportion of bad events that a bin must contain. Default is 0.05.}
}
\value{
A list with the following elements:
\itemize{
\item \code{feature_woe}: Numeric vector with the WoE assigned to each instance of the processed categorical variable.
\item \code{bin}: DataFrame with the generated bins, containing the following fields:
\itemize{
\item \code{bin}: Names of the categories grouped into each bin.
\item \code{woe}: Weight of Evidence (WoE) for each bin.
\item \code{iv}: Information Value (IV) for each bin.
\item \code{count}: Total number of observations in each bin.
\item \code{count_pos}: Count of positive events in each bin.
\item \code{count_neg}: Count of negative events in each bin.
}
\item \code{woe}: Numeric vector with the WoE for each bin.
\item \code{iv}: Total Information Value (IV) calculated for the variable.
\item \code{pos}: Vector with the count of positive events in each bin.
\item \code{neg}: Vector with the count of negative events in each bin.
}
}
\description{
This function processes a categorical variable by grouping rare categories, calculating event rates, and iteratively merging bins to maximize Information Value (IV) while ensuring monotonicity of event rates. It also calculates WoE and IV for the generated bins.
}
