% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OptimalBinningCategoricalWoE}
\alias{OptimalBinningCategoricalWoE}
\title{Optimal Binning and Weight of Evidence Calculation for Categorical Variables}
\usage{
OptimalBinningCategoricalWoE(
  dt,
  target,
  features,
  method,
  min_bins,
  max_bins,
  control,
  positive,
  preprocessed_data
)
}
\arguments{
\item{dt}{A data.table containing the dataset.}

\item{target}{The name of the target variable.}

\item{features}{Vector of categorical feature names to process.}

\item{method}{The binning method to use.}

\item{min_bins}{Minimum number of bins.}

\item{max_bins}{Maximum number of bins.}

\item{control}{A list of additional control parameters.}

\item{positive}{Character string specifying which category should be considered as positive.}

\item{preprocessed_data}{List of preprocessed data for each feature.}
}
\value{
A list containing results, reports, woebins, bestmod, and failed_features.
}
\description{
Optimal Binning and Weight of Evidence Calculation for Categorical Variables
}
