% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimalBinningGainsTable}
\alias{OptimalBinningGainsTable}
\title{Generates a Gains Table from the results of optimal binning}
\usage{
OptimalBinningGainsTable(binning_result)
}
\arguments{
\item{binning_result}{A list containing the binning results, which must include a data frame with the following columns: "bin", "count", "count_pos", "count_neg", and "woe".}
}
\value{
A data frame containing the following columns for each bin:
\itemize{
\item \code{bin}: The bin labels.
\item \code{count}: Total count of observations in the bin.
\item \code{pos}: Count of positive events in the bin.
\item \code{neg}: Count of negative events in the bin.
\item \code{woe}: Weight of Evidence (WoE) for the bin.
\item \code{iv}: Information Value (IV) contribution for the bin.
\item \code{total_iv}: Total Information Value (IV) across all bins.
\item \code{cum_pos}: Cumulative count of positive events up to the current bin.
\item \code{cum_neg}: Cumulative count of negative events up to the current bin.
\item \code{pos_rate}: Rate of positive events within the bin.
\item \code{neg_rate}: Rate of negative events within the bin.
\item \code{pos_perc}: Percentage of positive events relative to the total positive events.
\item \code{neg_perc}: Percentage of negative events relative to the total negative events.
\item \code{count_perc}: Percentage of total observations in the bin.
\item \code{cum_count_perc}: Cumulative percentage of observations up to the current bin.
\item \code{cum_pos_perc}: Cumulative percentage of positive events up to the current bin.
\item \code{cum_neg_perc}: Cumulative percentage of negative events up to the current bin.
\item \code{cum_pos_perc_total}: Cumulative percentage of positive events relative to total observations.
\item \code{cum_neg_perc_total}: Cumulative percentage of negative events relative to total observations.
\item \code{odds_pos}: Odds of positive events in the bin.
\item \code{odds_ratio}: Odds ratio of positive events compared to the total population.
\item \code{lift}: Lift of the bin, calculated as the ratio of the positive rate in the bin to the overall positive rate.
\item \code{ks}: Kolmogorov-Smirnov statistic, measuring the difference between cumulative positive and negative percentages.
\item \code{gini_contribution}: Contribution to the Gini coefficient for each bin.
}
}
\description{
This function takes the result of the optimal binning process and generates a detailed gains table. The gains table includes metrics such as the Weight of Evidence (WoE), Information Value (IV), cumulative positive and negative percentages, Kolmogorov-Smirnov (KS) statistic, odds ratio, lift, and Gini contribution for each bin.
}
\examples{
\dontrun{
binning_result <- OptimalBinning(target, feature)
gains_table <- OptimalBinningGainsTable(binning_result)
}

}
