% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OptimalBinningMapTargetVariable}
\alias{OptimalBinningMapTargetVariable}
\title{Map Target Variable}
\usage{
OptimalBinningMapTargetVariable(dt, target, positive)
}
\arguments{
\item{dt}{Data table}

\item{target}{Target variable name}

\item{positive}{Positive class indicator}
}
\value{
Updated data table with mapped target variable
}
\description{
Map Target Variable
}
\keyword{internal}
