% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptimalBinningNumericChiMerge}
\alias{OptimalBinningNumericChiMerge}
\title{Performs optimal binning of a numeric variable for Weight of Evidence (WoE) and Information Value (IV) using the ChiMerge algorithm}
\usage{
OptimalBinningNumericChiMerge(
  target,
  feature,
  min_bins = 2L,
  max_bins = 7L,
  pvalue_threshold = 0.05,
  bin_cutoff = 0.05,
  min_bads = 0.05,
  max_n_prebins = 20L
)
}
\arguments{
\item{target}{Integer vector representing the binary target variable, where 1 indicates a positive event (e.g., default) and 0 indicates a negative event (e.g., non-default).}

\item{feature}{Numeric vector representing the numeric variable to be binned.}

\item{min_bins}{(Optional) Minimum number of bins to generate. Default is 2.}

\item{max_bins}{(Optional) Maximum number of bins to generate. Default is 7.}

\item{pvalue_threshold}{(Optional) P-value threshold for the chi-square test used to determine whether to merge bins. Default is 0.05.}

\item{bin_cutoff}{(Optional) Cutoff value that determines the frequency of values to define pre-bins. Default is 0.05.}

\item{min_bads}{(Optional) Minimum proportion of bad events (positive target events) that a bin must contain. Default is 0.05.}

\item{max_n_prebins}{(Optional) Maximum number of pre-bins to consider before final binning. Default is 20.}
}
\value{
A list with the following elements:
\itemize{
\item \code{feature_woe}: Numeric vector with the WoE assigned to each instance of the processed numeric variable.
\item \code{bin}: DataFrame with the generated bins, containing the following fields:
\itemize{
\item \code{bin}: String representing the range of values for each bin.
\item \code{woe}: Weight of Evidence (WoE) for each bin.
\item \code{iv}: Information Value (IV) for each bin.
\item \code{count}: Total number of observations in each bin.
\item \code{count_pos}: Count of positive events in each bin.
\item \code{count_neg}: Count of negative events in each bin.
}
\item \code{woe}: Numeric vector with the WoE for each bin.
\item \code{iv}: Total Information Value (IV) calculated for the variable.
\item \code{pos}: Vector with the count of positive events in each bin.
\item \code{neg}: Vector with the count of negative events in each bin.
}
}
\description{
This function processes a numeric variable by removing missing values and creating pre-bins based on unique values. It iteratively merges bins using the Chi-square test of independence, with a p-value threshold to determine whether bins should be merged. It ensures the minimum number of bad events (\code{min_bads}) is respected, while also calculating WoE and IV for the generated bins.
}
