% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalBinningR.R
\name{OptimalBinningselectBestModel}
\alias{OptimalBinningselectBestModel}
\title{Select Best Binning Model}
\usage{
OptimalBinningselectBestModel(dt, target, feature, control, min_bins, max_bins)
}
\arguments{
\item{dt}{A data.table containing the dataset.}

\item{target}{The name of the target variable.}

\item{feature}{The name of the feature to bin.}

\item{control}{A list of additional control parameters.}

\item{min_bins}{Minimum number of bins.}

\item{max_bins}{Maximum number of bins.}
}
\value{
The binning result of the best method.
}
\description{
This function tests multiple binning methods and selects the best one based on Information Value (IV).
}
\keyword{internal}
