% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_obwoe.R
\name{prep.step_obwoe}
\alias{prep.step_obwoe}
\title{Prepare the Optimal Binning Step}
\usage{
\method{prep}{step_obwoe}(x, training, info = NULL, ...)
}
\arguments{
\item{x}{A step_obwoe object.}

\item{training}{A tibble or data frame containing the training data.}

\item{info}{A tibble with column metadata from the recipe.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A trained step_obwoe object with \code{binning_results} populated.
}
\description{
Fits the optimal binning models on training data. This method is called
by \code{\link[recipes]{prep}} and should not be invoked directly.
}
