% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{print.oblr}
\alias{print.oblr}
\title{Print Method for oblr Objects}
\usage{
\method{print}{oblr}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{oblr}.}

\item{digits}{Number of significant digits to display. Defaults to the maximum between 3 and \code{getOption("digits") - 3}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Prints a brief summary of the \code{oblr} model, including estimated coefficients and convergence information.
}
