% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_obwoe.R
\name{required_pkgs.step_obwoe}
\alias{required_pkgs.step_obwoe}
\title{Required Packages for step_obwoe}
\usage{
\method{required_pkgs}{step_obwoe}(x, ...)
}
\arguments{
\item{x}{A step_obwoe object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A character vector of package names.
}
\description{
Lists the packages required to execute the step_obwoe transformation.
}
