% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{step.default}
\alias{step.default}
\title{Step function}
\usage{
\method{step}{default}(object, ...)
}
\arguments{
\item{object}{obrl class fit}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Step function
}
