% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_obwoe.R
\name{tunable.step_obwoe}
\alias{tunable.step_obwoe}
\title{Tunable Parameters for step_obwoe}
\usage{
\method{tunable}{step_obwoe}(x, ...)
}
\arguments{
\item{x}{A step_obwoe object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A tibble describing tunable parameters.
}
\description{
Returns information about which parameters of step_obwoe can be tuned
using the \code{tune} package.
}
