% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{update.oblr}
\alias{update.oblr}
\title{Update Method for oblr Objects}
\usage{
\method{update}{oblr}(object, formula., data., ...)
}
\arguments{
\item{object}{An object of class \code{oblr}.}

\item{formula.}{A new formula for the model. If not specified, the original formula is retained.}

\item{data.}{New data for fitting the model. If not specified, the original data is retained.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A new object of class \code{oblr} fitted with the updated parameters.
}
\description{
Updates the \code{oblr} model with new parameters without refitting the entire model.
}
