% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblr.R
\name{vcov.oblr}
\alias{vcov.oblr}
\title{Variance-Covariance Matrix Method for oblr Objects}
\usage{
\method{vcov}{oblr}(object, ...)
}
\arguments{
\item{object}{An object of class \code{oblr}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A variance-covariance matrix of the estimated coefficients.
}
\description{
Returns the variance-covariance matrix of the estimated coefficients from the \code{oblr} model.
}
