% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{.dispatch_algorithm}
\alias{.dispatch_algorithm}
\title{Internal Algorithm Dispatcher}
\usage{
.dispatch_algorithm(
  feat_type,
  algorithm,
  target_vec,
  feat_vec,
  min_bins,
  max_bins,
  control
)
}
\description{
Internal Algorithm Dispatcher
}
\keyword{internal}
