% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe_report.R
\name{.ob_cutoff_table}
\alias{.ob_cutoff_table}
\title{Internal: Cutoff Strategy Table}
\usage{
.ob_cutoff_table(score, y, n_points = 20L)
}
\description{
Internal: Cutoff Strategy Table
}
\keyword{internal}
