% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OBCalculateSpecialWoE}
\alias{OBCalculateSpecialWoE}
\title{Calculate Special WoE for Edge Cases}
\usage{
OBCalculateSpecialWoE(target)
}
\arguments{
\item{target}{Numeric or factor vector of two values.}
}
\value{
Numeric WoE.
}
\description{
Calculate Special WoE for Edge Cases
}
