% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OBGetAlgoName}
\alias{OBGetAlgoName}
\title{Get Available Optimal Binning Algorithms}
\usage{
OBGetAlgoName()
}
\value{
List with $char (categorical) and $num (numerical) named lists.
}
\description{
Get Available Optimal Binning Algorithms
}
