% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OBSelectAlgorithm}
\alias{OBSelectAlgorithm}
\title{Select Optimal Binning Algorithm}
\usage{
OBSelectAlgorithm(feature, method, dt, min_bin, max_bin, control)
}
\arguments{
\item{feature}{The name of the feature to bin.}

\item{method}{The binning method to use.}

\item{dt}{A data.table containing the dataset.}

\item{min_bin}{Minimum number of bins.}

\item{max_bin}{Maximum number of bins.}

\item{control}{A list of additional control parameters.}
}
\value{
A list containing the selected algorithm, its parameters, and the method name.
}
\description{
This function selects the appropriate binning algorithm based on the method and variable type.
}
\keyword{internal}
