% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{obwoe}
\alias{obwoe}
\title{Optimal Binning and Weight of Evidence Calculation}
\usage{
obwoe(
  dt,
  target,
  features = NULL,
  method = "auto",
  preprocess = TRUE,
  min_bins = 3,
  max_bins = 4,
  control = list(),
  positive = "bad|1"
)
}
\arguments{
\item{dt}{A data.table containing the dataset.}

\item{target}{The name of the target variable (must be binary).}

\item{features}{Vector of feature names to process. If NULL, all features except the target will be processed.}

\item{method}{The binning method to use. Can be "auto" or a specific method name.}

\item{preprocess}{Logical. Whether to preprocess the data before binning.}

\item{min_bins}{Minimum number of bins.}

\item{max_bins}{Maximum number of bins.}

\item{control}{A list of additional control parameters.}

\item{positive}{Character string specifying which category should be considered as positive. Must be either "bad|1" or "good|1".}
}
\value{
A list containing:
\item{woedt}{The original dataset with added WoE columns}
\item{woebins}{Information about the bins created}
\item{prepreport}{Preprocessing report for each feature}
\item{bestsreport}{Report on the best models used}
\item{failedfeatures}{List of features that failed processing}
\item{bestmethod}{Best method used for binning}
}
\description{
This function performs optimal binning and calculates Weight of Evidence (WoE)
for both numerical and categorical features.
}
