% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OBWoEMonotonic}
\alias{OBWoEMonotonic}
\title{Check if WoE values are monotonic}
\usage{
OBWoEMonotonic(woe_values)
}
\arguments{
\item{woe_values}{Numeric vector of WoE.}
}
\value{
TRUE if entirely nondecreasing or nonincreasing.
}
\description{
Check if WoE values are monotonic
}
