% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimal_binning_numerical_mdlp}
\alias{optimal_binning_numerical_mdlp}
\title{Optimal Binning for Numerical Features using MDLP}
\usage{
optimal_binning_numerical_mdlp(
  target,
  feature,
  min_bins = 3L,
  max_bins = 5L,
  bin_cutoff = 0.05,
  max_n_prebins = 20L,
  convergence_threshold = 1e-06,
  max_iterations = 1000L
)
}
\arguments{
\item{target}{Vetor inteiro binário (0 ou 1).}

\item{feature}{Vetor numérico da feature.}

\item{min_bins}{Número mínimo de bins (default: 3).}

\item{max_bins}{Número máximo de bins (default: 5).}

\item{bin_cutoff}{Proporção mínima de registros por bin (default: 0.05).}

\item{max_n_prebins}{Número máximo de pré-bins (default: 20).}

\item{convergence_threshold}{Limite de convergência (default: 1e-6).}

\item{max_iterations}{Número máximo de iterações (default: 1000).}
}
\value{
Uma lista com bins, woe, iv, contagens, cutpoints, convergência e iterações.
}
\description{
Executa binning ótimo usando o Princípio do Comprimento Mínimo da Descrição (MDLP).
Cria bins de modo a minimizar a perda de informação, mesclando bins adjacentes que reduzem o custo MDL,
e assegurando monotonicidade no WoE. Ajusta o número de bins entre min_bins e max_bins, mesclando bins raros.
}
