% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OptimalBinningPreprocessData}
\alias{OptimalBinningPreprocessData}
\title{Preprocess Data for Optimal Binning}
\usage{
OptimalBinningPreprocessData(dt, features, control)
}
\arguments{
\item{dt}{A data.table containing the dataset.}

\item{features}{Vector of feature names to process.}

\item{control}{A list of control parameters.}
}
\value{
A list of preprocessed data for each feature.
}
\description{
Preprocess Data for Optimal Binning
}
\keyword{internal}
