% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalBinningR.R
\name{OptimalBinningSelectBestModel}
\alias{OptimalBinningSelectBestModel}
\title{Select Best Binning Model}
\usage{
OptimalBinningSelectBestModel(dt, target, feature, min_bin, max_bin, control)
}
\arguments{
\item{dt}{A data.table containing the dataset.}

\item{target}{The name of the target variable.}

\item{feature}{The name of the feature to bin.}

\item{min_bin}{Minimum number of bins.}

\item{max_bin}{Maximum number of bins.}

\item{control}{A list of additional control parameters.}
}
\value{
A list containing the best binning result and additional information.
}
\description{
This function selects the best binning model by testing multiple methods and comparing their performance.
}
\keyword{internal}
