% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwoe.R
\name{OptimalBinningSelectBestModel}
\alias{OptimalBinningSelectBestModel}
\title{Select Best Binning Model}
\usage{
OptimalBinningSelectBestModel(
  dt_binning,
  target,
  feature,
  min_bins,
  max_bins,
  control,
  allowed_methods,
  trace = FALSE
)
}
\arguments{
\item{dt_binning}{Data for binning}

\item{target}{Target variable name}

\item{feature}{Feature variable name}

\item{min_bins}{Minimum number of bins}

\item{max_bins}{Maximum number of bins}

\item{control}{Control parameters}

\item{allowed_methods}{Vector of allowed binning methods}

\item{trace}{Logical, whether to generate error logs when testing existing methods}
}
\value{
Best binning result
}
\description{
Select Best Binning Model
}
\keyword{internal}
