% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalBinningR.R
\name{OptimalBinningValidateInputs}
\alias{OptimalBinningValidateInputs}
\title{Validate Inputs for Optimal Binning}
\usage{
OptimalBinningValidateInputs(
  dt,
  target,
  feature,
  method,
  min_bins,
  max_bins,
  cat_cutoff,
  bin_cutoff,
  control
)
}
\arguments{
\item{dt}{A data.table containing the dataset.}

\item{target}{The name of the target variable.}

\item{feature}{Optional. Name of a specific feature to process.}

\item{method}{The binning method to use.}

\item{min_bins}{Minimum number of bins.}

\item{max_bins}{Maximum number of bins.}

\item{cat_cutoff}{Cutoff for categorical variables.}

\item{bin_cutoff}{Cutoff for numeric variables.}

\item{control}{A list of additional control parameters.}
}
\value{
No return value, called for side effects
}
\description{
This function validates all inputs for the OptimalBinningWoE function.
}
\keyword{internal}
